refactor : 재고 수량 타입 변경에 따른 DTO 타입 수정

+++ product/src/main/java/com/hubEleven/stock/application/dto/StockResult.java
@@ -5,14 +5,16 @@ import java.util.UUID;
 
 // Service 계층의 반환용 DTO (Domain 엔티티와 API 응답 사이의 중간 계층)
 public record StockResult(
-		UUID stockId, UUID productId, String productName, UUID hubId, UUID companyId, Long quantity) {
-	public static StockResult from(Stock stock, String productName) {
-		return new StockResult(
-				stock.getStockId(),
-				stock.getProductId(),
-				productName, // productName 포함
-				stock.getHubId(),
-				stock.getCompanyId(),
-				stock.getQuantity());
-	}
+        UUID stockId, UUID productId, String productName, UUID hubId, UUID companyId,
+        int quantity) {
+
+    public static StockResult from(Stock stock, String productName) {
+        return new StockResult(
+                stock.getStockId(),
+                stock.getProductId(),
+                productName, // productName 포함
+                stock.getHubId(),
+                stock.getCompanyId(),
+                stock.getQuantity());
+    }
 }
